feat(institutp-sec): Se trabaja en la securización, sigue sin ir confirmación. Ya están la mitad de los test

+++ 2oTrimestre/tareas-clase/instituto-sec/src/main/java/es/iespuertodelacruz/pme/institutosec/dto/usuario/UsuarioDTOSalidaV3.java
@@ -7,7 +7,7 @@ public record UsuarioDTOSalidaV3(
         String nombre,
         String correo,
         String rol,
-        byte verificado,
+        int verificado,
         String tokenVerificacion,
         Date fechaCreacion
